Separating various chunks.
